Logwatch: Capture last run time before starting logwatch
The new last run time is captured before logwatch is run so that
there is no possibility for a gap after the time the logwatch process
is started and the last run time is recorded. (Because such a gap
would theoretically not appear in the logs. Practically, Logwatch
should round the time up anyways.)